Always show auxiliary subtypes from NavBar keyboard icon.
The auxiliary subtypes should be listed if the input method picker is
opened from NavBar keyboard icon.  However there is only
IMM#showInputMethodPicker() API to open input method picker and this is
also used from LockScreen or Settings UI.  Auxiliary subtypes should not
be listed there(Id7cf5d122).  Thus framework shows auxiliary subtypes
based on IMMS#mInputShown and LockScreen state, but it is not a perfect
solution.  If a physical keyboard is connected, the soft input may be
gone.  As the result, auxiliary subtypes won't be listed even if it is
opened from NavBar keyboard icon.

To fix this issue, this CL introduces IMM#showInputMethodPicker(boolean)
to be able to decide showing auxiliary subtypes by caller.
Note that IMM#showInputMethodPicker(boolean) is still hidden with @hide.
There is no public API change in this CL.

Bug: 20763994

Change-Id: I1e50ee42838a1bf64a612da4904aa93458d44ea4

Former-commit-id: 80ff4ed6bb8dbdad7192d679a01096aa888e090b