Fix checkbox animation end colors
Uses a themed color state list so that even though the animations will
end showing the same drawable that they started from, and therefore the
same tint, the correct color will be displayed based on the state. Also
renames the checkbox assets so it's easier to figure out which one is
going in which direction.

Bug: 20733852
Change-Id: I1e8331d9f035595953db6df3b28b766db3617e1c

Former-commit-id: e318a467e103e2d08b788f0f99c9e6672e571692