When returning, transition only existing shared elements.
Previously all shared element names were delivered when
returning from an Activity, even if no shared elements
existed in the exiting Activity. That prevented the calling
Activity from showing a shared element that it delivered,
but will not be receiving back again until shared elements
are transfered. That leaves a weird hole in the UI.

Now only the shared elements mapped in the returning view
hierarchy are delivered to the calling Activity.

Change-Id: I481a8bc7a771d7e819871f424d22313314c9ee8b

Former-commit-id: 8de1de1f0450c25ce6079f4b8674d4bb0114287e