TIF: Make all the parameters for UNRATED null
UNRATED is equivalent to Java null in its meaning and usage, thus it is
not appropriate to use non-trivial strings such as "com.android.tv" and
"UNRATED" to create this constant. Replaced them with the string "null",
which is commonly used in other parts of Android and also modified the
documentation to clarify its usage.

Bug: 19102311
Change-Id: I28ac8da3584eb2cc6dfd7284e2b88873fd4e9d13

Former-commit-id: 936c040ec445afad98ec16fc634ae6573eceefbb