Fixed a bug where the HUN would interfere with the normal touches
When a hun came in or went away, the touchable regions were not
enforced correctly due to a race condition. This lead to the user
grabing the notification panel unintentionally and ripping him
out of the normal touch state.
Also fixed a small bug where the alpha was not correctly animating
sometimes and the HUN shadow would still draw.

Bug: 20956211
Change-Id: Iae1fef5825b3d2b8b4128cc8c3272019194cd819

Former-commit-id: 8b4a06e9e269140c93c1a9ef5add008f7610d1a4