New custom widgets library
Adds the widgets needed by the theme editor as a jar file that will be
linked from AS. The controls currently are:
- ThemePreviewLayout: A custom layout that allows to dynamically reflow
  cards that display the theme preview.
- PressedButton: Custom button to display in a pressed state (currently
  we can only do it programatically).
- ErrorCatcher: View that can wrap other view and stops a view throwing
  exceptions that will break the theme editor rendering. This is
  currently used to wrap the custom controls loaded from the source
  code.

The new library currently depends on bridge to be able to use the logger
for the ErrorCatcher view. The ErrorCatcher view is a temporary solution
that will be replaced in a future CL and the dependency will be removed.

Change-Id: I832c447e22e1381abff74c46c4282921b3f6fa23

Former-commit-id: aa3b2043f0403d064afcf0977f6e713662aceaff