Postpone AnimatedVectorDrawable animator inflation until applyTheme()
This CL works around Animator's lack of support for applying a theme
after inflation by postponing Animator inflation until a theme is
available, either in inflate() or applyTheme().

Includes a workaround for AVDs that don't reference any theme attrs
in their animators and this don't require a call to applyTheme().

We'll follow up with real support for applyTheme() in Animator, at which
point we can remove this workaround.

Bug: 20817800
Change-Id: I5a378a76e3625b9d754cb74ae98c07ba7c5698e5

Former-commit-id: 9b115a9870a184e32bdbd07f792f9b8c956c75ca