Limit large font accessibility setting to user profile.
There were 2 problems:
1) When Settings application updated font size it called
ActivityManagerNative.updatePersistentConfiguration() that makes IPC
call to system_process. Then font size was persisted in system_process.
But, the user id we save with the font size if calculated based on the
current process, which is UserHandle.USER_OWNER for the system_process.

2) When user was changed font size was not read from database. Font size
from current configuration was always used.

b/18305168

Change-Id: Id847935a1ab1da3ef133e28ed6928c99c5fe0f16

Former-commit-id: 05075305b4bac26df1c97af1901171ba26f0a48b