docs: Quick fixes to screenshot
I'd been using the "srcset" tag wrong (we're updating the devsite
style guide) so I'm fixing that. Also mislabeled this as "Figure 2"
(it's currently the only figure on the page).

Since this is a very straightforward fix I'm going to go ahead and
submit it.

See first comment for doc stage location.

Change-Id: I81969ff1a3838e516ef00612f3ae8944cca8fab5

Former-commit-id: 1521c51cb69e73de2d37a98c367e5fd617056af6