Input command can fail (with exceptions) if the touchscreen id is not 0
We had a device where the touchscreen was reconfigured during system
init, and as a result was assigned a device id that was not 0.
The following message from InputReader.cpp was logged when this occured:
I/InputReader(  762): Device reconfigured: id=4, name='himax-touchscreen', size 1024x600, orientation 1, mode 1, display id 0

Attempting to use the input command for the touchscreen in this state
will fail due to the hardcoded touchscreen device id.

Change-Id: I4e8fad864e7b83cce00c680b254bad06e1200f15

Former-commit-id: 30fefa4d5df1d690f69c2db0b631565acc98b91c