Fix calls to Rect.intersect(Rect)
This CL checks for the return value for Rect.intersect(Rect) for whether
there is actually an intersection before taking the calling rect as the
intersection. In the case of no intersection (Rect.intersect(Rect) returns
false), the calling rect would not have been changed, so here in this CL
it will be manually set empty.

Bug: 7368679
Change-Id: I7494c1a1649eca09e34412f0a59b795e8b3a82aa

Former-commit-id: 942bc58518a82bc3dbeffe4d64f9a5d2d2d8cf43