Change to hide BEARER_BITMASK.
BEARER_BITMASK uses ServiceState constants that are not public so
keeping it hidden for now. Ideally it should use TelephonyManager
constants instead as they are already public. However the constant
values for same RATs are different between the two. So mapping from
one to the other is as apns xml files will also need to be updated
(including those of OEMs).

Bug: 21342148
Change-Id: Iced28ede187d3e72b3f969bc99af4edf34d86a36

Former-commit-id: 3ed641f65c3060205e304e8a0d13e88cc72d0af4