Respect accessibility importance for before/after node ordering.
One can set a view to be visited before/after another one for
accessibility but the before/after target views were always added
to the node info for the view regardless if they are included
for accessiilbity. Now the target viwes are reported only if they
should be included for accessibility.

Change-Id: Iba64e6f0ee0dc090c8c3e6a0a940df8c7ecd49d9

Former-commit-id: 2dca7ca9bfcb120f41d16923172796bb540103be