Move NotoSansSymbols higher than CJK in the fallback chain.
This makes sure that symbol characters who have both an alternative
or "full-width" version in the CJK fonts and a more regular version
in the symbols font get the default or "half-width" version by
default in non-CJK locales.

This is not a complete fix: such "full-width" symbol characters only
exist in the Simplfied Chinese font for now, which means in
Traditional Chinese, Japanese, and Korean locales, they characters
would now appear "half-width", which *may* be sub-optimal. The
complete fix would make sure the ones that actually need to be
"full-width" in TC/J/K are also included in the TC/J/K font.

Bug: 21525318
Bug: 21525012
Bug: 21524421
Bug: 21521330
Bug: 21520316
Bug: 21518544
Bug: 21517912
Bug: 21503807
Bug: 21502245
Bug: 21501669
Bug: 21525905
Bug: 21524018
Bug: 21503952
Bug: 21502678
Change-Id: I93cf7119b45ed3c080cd39d631a687e0ef9fd8ce

Former-commit-id: 912dbdb9fa3558a5cff2c0c5d410a2abc978422b