Pass an args bundle to launchAssistAction.
In order to track the input device that was used to trigger assist, the
input device id is sent as an extra in the assist intent whenever it is
available. This is particularly useful on TVs, when an app may want to
know whether the input device has a microphone.

Bug: 21666123
Change-Id: I0f8c09e2f617606bef481bdff924cb6b9b47dd12

Former-commit-id: 0e5f110fc915a9b044eb04cd07ae7ac588eacc8f