Include launched packageName in systrace.
This makes it easier to tell which package the activity manager is
launching. The downside here is that each launch shows up in a separate
row in the systrace, but this matches up with the "Start proc:" label,
which contains the package name as well and makes it easier to inspect
both simultaneously.

bug: 21632700
Change-Id: I20ebc2f2a2cf0dc74c6d7daa4bafa381ae4c4060

Former-commit-id: 7829c81345ce8b7e44ab22bf4433be77e0ebd8ae