Specifies the new height/width for update() on ListPopupWindow.show().
The current code is doing setHeight()/setWidth() and then invokes
update() with -1 for height/width params. However, javadoc for
setHeight() says that "If the popup is showing, calling this method
will take effect the next time the popup is shown", and update()
says "Width and height can be set to -1 to update location only."

That said, ListPopupWindow should invoke update() with the obtained
width/height instead of invoking setWidth()/setHeight() to resize
the popup window correctly.

Bug: 21276583
Change-Id: I0961820ff2e845a0f68c20d6eea827aa898c2101

Former-commit-id: 0eaec1f5d1b47739b8d33cb45111bccd18f2d7b1