DO NOT MERGE ANYWHERE Revert "Always re-establish kernel alarms when considering new alarm set"
This reverts commit 926225855b69f05c8a626d69652430c2e7a153da [formerly c83d3e4b72f8f25bbac39a577f125f0ca04354da].

While the above patch does help deal with the issue of "lost" kernel
alarms, experiments found it to also cause the unnecessary duplicate
setting of non-lost alarms in other devices.

This leads to a situation where the CPU constantly tries to sleep, but
occasionally keeps failing to do so, because of a collection of these
duplicate alarms constantly activating the "alarm" kernel wakeup source
, presumably due to them being triggered.

These might have accumulated over several iterations when the kernel
alarms were set again (via rescheduleKernelAlarmsLocked()).

To prevent the CPU thrashing (which heavily drains power on more
battery-constrained devices, and increases the number of alarms being
set), the aforementioned commit is backed out.

Bug: 20136729
Change-Id: I2a189cb303deb78a2faf476b0e789445407f3e11

Former-commit-id: d19ffd4f2ca4130c41dd5f6baddad1830598ecbc