Delete TODO and file a bug.
This TODO looks more like a feature request, so I filed b/21766386 to
evaluate the feature. We should remove the TODO especially because this
text is in the public API.

Note: I'm not deleting the exact same comment from
packages/services/Telephony/res/values/config.xml
because the data migrated to CarrierConfigManager is slated to be
deleted from config.xml files.

Bug: 21571894
Change-Id: Iaba042fb47ade24f5825a48b21960193d322363f

Former-commit-id: 5bf0ca1b3a304faf265a6c3e63410b32f4fdb118