Only read foreground attr for FrameLayouts for targetSdkVersion < M
Now that the android:foreground attribute is supported for all views
it turns out some existing apps were previously setting foreground or
applying styles with a foreground to unsupporting views even though
the values being set were not drawables. While this was silently
ignored before, now it results in app crashes.

The most expedient fix is to only support foreground on
non-FrameLayout views if the app sets targetSdkVersion to M or newer.

Bug 21646309

Change-Id: I8937f44988866e86e640555f4bbc5dd5b2e1cce1

Former-commit-id: bbd3eeb9e53707bcbf0cc540acd7bb566252a969