Fix IME window flickering during rotation.
This was introduced by Ifd15736b163ab,
performLayoutAndPlaceSurfacesLocked is called even if
computeNewConfigurationLocked() returns non-null object.
This is simply by mistake and now computeNewConfigurationLocked
never returns null. The only case we need to care is that
mDisplayReady is false, but there is nothing to do with that state.

Thus simply removes if segments from computeNewConfiguration.

Bug: 20823978

Change-Id: I527dfeddffb8d928d578f8d60d64f98557aa3dcb

Former-commit-id: d56ec1d44479cac0820c6afe9aaa689fc47c431d