Trace Process.killProcessGroup calls in server/am/
to make it easier to find boot time regressions like 21813611
and 21861473.

Process.killProcessGroup is a blocking call and usually takes
about 50ms to finish on a recent device even when the process
group contains only one process. Since Process.killProcessGroup
tends to be called in a loop, the function can be a bottleneck.

Bug: 21813611
Change-Id: I9c1727c0c0b321a30c96321198cc02f438496ded

Former-commit-id: cf4b2542fae9d6c5d3788c6a935824094fb1c1ea