Fixes for Region_writeToParcel.
Check the return value of Parcel::writeInplace. If it is NULL, there
was a failure, so do not attempt to write to it. Instead, report the
error and return false.

If SkRegion::writeToMemory claims to have written a different amount of
memory than it claimed it needed, report that error as well.

Change uses of NULL in this function to nullptr.

BUG:21271229
BUG:20666821
Change-Id: Ia6160f74f30bf42f5ff97f716dadb01d1f0d6961

Former-commit-id: 2a6d6e504712dcddd030a6007d9dfbb089258619