Migrate 'csp_enabled' to CarrierConfigManager.
This variable controls whether SIM cards are allowed to disable manual
carrier selection, so it makes more sense for it to be managed as
carrier config.

Currently the variable is only used in device overlays
for specific networks, but we are dropping the device-overlay so that
this simply applies to those networks.

Change-Id: Ia6ba6a89c94c0916c19bdf792e48a4feac87eded

Former-commit-id: 019523a8a44dcab0bac7cb4fa3c22ada7cca6804