Remove not needed contacts related perissions.
Removing the read/write profile/social stream permissions as they
are not needed anymore. These permissions are for accessing data
nested in the conacts provider which is already guaded by the
read/write contacts runtime permissions. The removed permissions
would be in the contacts group which means they would not provide
more protection compated to read/write contacts. Also removing
the permissions voids the need for app op support for legacy apps.

Removed deprecated APIs for social streams as these were deprecated
and will go away in the next release. We want apps targeting M to
not be able to compile if still suing these APIs to help with
migration.

Change-Id: I26ed9055847af7f92c78eb0f4ac8f9f1aa616fcd

Former-commit-id: a2991da0d671adf511ccb884cf25bf1241303f92