don't clear notification effects in response to peeks
When the heads up appears it generates a visibility event for the panel.
This always clears the notification effects (buzz, beep, blink).
If we suppress that for heads up notifications, we never get another chance.
So we add a new call when we open the shade from the heads up to catch this case.

Bug: 20947591
Change-Id: I233e5346617ca2fce7851bd85154c838849df8f7

Former-commit-id: 1689594851a6dd985cff69913e7f4bc548062808