Fixed a bug with encryption where the keyguard was never shown with PIN
This led to the network being unusable.
We now ensure that the Keyguard is not notified of the SIM PIN
changes until after the initial dismissal logic is done.
In addition, the fullscreen PIN is now always shown immediately
and not until the user swipes up, which was an error before.
Also made sure that the Keyguard Scrim is correctly visible
when shown.

Bug: 21850021
Change-Id: I81804124bae30adf0e17e26cc42f67e48fc446e8

Former-commit-id: edd32b881780e52b4ecc43f8df3b6b091e70a863