Preventing leak
When transitioning from ON to OFF with LE Advertisers, advertiser do not
get a chance to unregister themselves as the stopAdvertising checks the
state of the stack and throws before unregistering the object.

It will then never remove the callback objects causing a leak.

b/22092678 | Remote service crash after switching to restricted profile

Change-Id: I04817026a524d10d60abdd8b533554a71a0112e2

Former-commit-id: cd62b10bd612e0c49083cc03f4ac63376a52fe8c