animateNoDamage failed to remove from current frame list
runRemainingAnimations is used to animate the available animation handle
which were not yet run, it actually animate those animation handle and
removes from the list by notifyAnimationsRan(). In case if these animation
handles are not removed, it leads to the Fatal Assert of runRemainingAnimations.

Some usecases causes the addition of animation handle when there is
no animators, in these cases this check prevents execution of
notifyAnimationsRan() leads to Fatal Assert crash of runRemainingAnimations.

Allowing the animation to happen irrespective of animators solves the problem

Change-Id: I685dc7ebf736268927a536595910fc63353306eb
Signed-off-by: Parasuraman <parasur@motorola.com>
Signed-off-by: Fred Fettinger <fettinge@motorola.com>

Former-commit-id: 2ccb5030c691849c1dfadf9c7136b7fc18acab9c